+++ .github/CHANGELOG.md
@@ -39,6 +39,7 @@
 1. [HYD] Fixed too slow leak measurement valves operation - @Crocket63 (crocket)
 1. [ISIS] Added temporary ISIS font with arrows - @aweissoertel (Alexibexi#7550)
 1. [PFD] Improve PFD barberpole rendering and behaviour - @lukecologne (luke)
+1. [HYD] Fixed actuator position control demand consistency - @Crocket63 (crocket)
 1. [FMGC] Only emit decel point when an approach is selected - @tracernz (Mike)
 1. [HYD] Fixed SFCC computer failing to send FPPU commands - @Crocket63 (crocket)
 1. [MCDU] Fix padding of arc radii on F-PLN - @tracernz (Mike)